Change primary and dark theme colors in mint.json
Set 'primary' & 'dark' key values under 'colors' in mint.json from '#141414' to '#121212' and 'background' 'dark' from '#ffffff' to '#0A0A0A'. In the 'anchors' object 'from' & 'to' were updated from '#141414' to '#121212' and '#181818' respectively. These changes are implemented to enhance the overall look and improve the visibility of the dark theme based on the user feedback received.