Add Android WiFi instructions, update iOS wording
Added a new page providing detailed instructions on how to use Illusory with Android's Wi-Fi settings. Added reference to this page in the iOS group of the mint.json file for easier navigation. Also, made a minor copy update in the iOS Wi-Fi app documentation for better clarity on the most popular apps used.