Refactor documentation by renaming files for clarity
This commit streamlines organizational clarity in the documentation. We've renamed the guidance files for setting up on macOS and Mozilla Firefox to make them more intuitive to locate. We believe this restructuring will improve the ease of navigation and understanding for users, helping them incorporate Illusory across different platforms more efficiently.