Remove "Documentation" from anchor list in mint.json
"Documentation" anchor with its details has been removed from the mint.json file. The removal is in line with the ongoing restructuring of data to enhance the user experience. The "Documentation" link has been deprecated and its content will be merged with "API Reference" to provide more consolidated and intuitive navigation for the users.