Update API Reference URL in mint.json
This commit addresses a necessary update to the URL for the "API Reference" section in the mint.json file. Previously, the URL was set as "api-reference/introduction", however, the correct location of the document is under "docs/api-reference/introduction". This change should now correctly direct users to the respective documentation section.