Refactor schema in OpenAPI specification
Changed the 'Plant' schema in the OpenAPI documentation (openapi.json) to 'Proxy'. This changelog encompasses modification in the schema object as well. The update involved defining 'ok' as a 'boolean' and 'data' - an object with a far more detailed specification as compared to the 'Plant' schema. The example properties for both 'ok' and 'data' were added for developer's understanding. This was done to reflect recent updates in the API endpoints and to provide appropriate information to client developers.