Refactor proxy-list detail structure in openapi.json
Simplified the schema of the proxy-list in openapi.json file. Instead of describing all properties of proxy-list item, the schema was changed to refer to a pre-defined model `proxy-get`. This change makes the API responses more consistent by using the same model for single and multiple proxy responses. The modification improves maintainability and extensibility of the code as any changes to the model will be reflected in all responses containing it.