Update schema in API documentation
Updated the schema in the api-reference/openapi.json file. The 'ok' field which indicated the success of a request was removed from proxy-get and proxy-list definitions. This adjustment was made as the success of a request should be indicated by the HTTP status code of the response and not in the response body. This simplifies the API response and aligns it more with standard APIs using HTTP protocol. It also reduces unnecessary complexity in client-side code that consumes this API.