Remove object type from 'data' in openapi.json
The object type and its description were removed from the 'data' field in the openapi.json file. This change was done to correctly indicate that 'data' is used to refer to an array of proxy details (specified under the 'proxy-get' schema) and not a singular object as previously implied, aligning it with the actual server response when handling multiple proxies.