Add bearerAuth mechanism to openapi.json
A `bearerAuth` mechanism is introduced in the API structure in this commit for enhanced security. The previous requirement of `apiKey` in the URL for fetching particular proxies is now replaced with this new method. Descriptions are also updated for better clarity and a new route for changing the external mobile IP is added.