Update API documentation, remove apiKey from path
The API documentation and JSON schema have been updated to remove the apiKey from being part of the path parameters, and instead use bearer token for authentication. This change is to enforce better security practices and streamline usage of endpoints. The descriptions of certain endpoint responses have also been clarified to be more informative.