Refine descriptions for success and failure conditions in API
The change adds concrete examples to the descriptions of success and failure conditions in our API reference to provide more clarity. The modifications specify that a successful request will return 'true' and an unsuccessful one will return 'false'. The goal is to provide better guidance to developers for interpreting these Boolean response attributes.