Update JSON configuration for API docs
Updated the mint.json file to include a reference to the new "proxy-change-ip-link" endpoint in the API documentation. This is to reflect the addition of the new functionality available for changing IP through a link. This provides users with a more precise description of how to use the limited API key alongside the proxy name to change IP effectively. Additionally updated the `post` method of the `/v1/proxies/changeip/{proxyName}` endpoint to improve accuracy in its description.