Add proxy basic auth endpoint to API docs
Documentation has been expanded to include the new endpoint `proxy-auth-basic`, a feature that allows updating basic authorization of a proxy. This commit includes the addition of its associated mdx file and modifications to `openapi.json` and `mint.json` to reflect this change. By documenting this update, users of this API will be better guided on its usage and implementation.