Update endpoint description in API docs
Modified the description of the `proxy-auth-basic` endpoint in the `openapi.json` file. The previous "testing summary" text and generic description were replaced with a clear warning message to provide better guidance to users. This is part of ongoing efforts to improve the documentation of our API for easier use and implementation.