Modify title and description in proxy-auth-basic.mdx and openapi.json
Changes have been made to documentation of the `proxy-auth-basic` endpoint. In the .mdx file, the title was simplified to 'Basic Authentication' and a description for 'proxyName' parameter was added. Meanwhile, the 'Test Warning' message in openapi.json was replaced with a clearer, more descriptive explanation of what the 'PATCH' request does. These edits provide more clarity to users of our API.