Update docstrings for proxy authorization
This commit enhances the clarity of the documentation related to proxy authorization methods - `Basic Authentication` and `Whitelist Authentication`. The updates provide a more precise explanation of how proxies deal with usernames, passwords and whitelisted IPs upon deployment by default. Furthermore, the description in the openapi schema for the patch method of the whitelisted IPs endpoint has been clarified to depict the purpose as setting the authorized IPs for a proxy server. This provides a better understanding to users interfacing with the API.