Update IPv6 test link format in proxy-ipv4.mdx
This commit modifies the link format in the warning message of 'proxy-ipv4.mdx'. The hyperlink previously incorporated in the text body has been replaced with HTML tags to offer a more readable and accessible experience. This change makes it easier for users to interpret and understand the important instructions provided for IPv6 testing.