Add reboot endpoint to proxy API documentation
Added a new POST reboot endpoint under the proxy section of the API reference. This endpoint allows users to hard reboot a proxy server. It takes the proxyName as a parameter and returns a server response message and data relating to the last proxy reboot. Documented this change in a new .mdx file, proxy-power-cycle.mdx. Updated file path in mint.json to include this new documentation. This change allows users to solve issues with their proxy server by facilitating easier reboot.