Add 422 error response to OpenAPI documentation
Extended OpenAPI documentation to include the 422 error in response to a request made to '/v1/proxies/swap/{proxyName}'. This was added to present the error scenario when the 'proxy-swap' request is unsuccessful. The 422 error response provides detailed information about the issue including a unique identifier for the proxy server and an error message. In addition, the 'proxy-swap' endpoint has been added to the 'mint.json' file, allowing the new endpoint to be visible in the table of contents.