Update response status codes in OpenAPI documentation
Changed several instances of HTTP 200 response code to 202 in the OpenAPI documentation. The change is in line with the implementation where certain API calls are designed to send a 202 accepted status instead of the usual 200 OK status to indicate that request has been accepted for processing, but the processing is not complete.