Update status code in openAPI docs to reflect actual functionality
The HTTP response status code in the OpenAPI documentation is updated from 200 to 202. This is to accurately reflect our application's behaviour, wherein specific API requests return a 202 status code. The 202 status is used when a request is accepted for processing, but it is not completed yet, differing from the usual 200 OK status.