Refine API proxy allocation documentation
This commit corrects redundant instructions in the 'POST /v1/proxies/allocate' endpoint documentation which was unnecessarily repeated. Now, it succinctly describes the role of `period` and `interval` parameters in extending a proxy's expiration date. It also improves readability of the warning related to the Illusory Wallet fund deduction when allocating time to a proxy.