Added Get Funds endpoint to API docs
The /v1/funds endpoint was added to the API reference documentation. This endpoint allows users to retrieve the balance of their Illusory Wallet. A new document was also created in the 'platform' group on 'Get Funds' to provide detailed information about this new API. This addition enhance our documentation with up-to-date features.