Refined descriptions and added new API reference in documentation
The descriptions for "upcoming" and "expiring" description fields in 'openapi.json' have been refined to specify charges expected in the next 24 hours. Additionally, a new API reference for 'Get Price' was added into 'price-get.mdx' and 'mint.json' files. The enhancements are made to provide more precise information and improve overall clarity for the users of the API.