Simplify pricing periods in API schema
In response to feedback about the complexity and rigidness of our API when working with pricing periods, the 'weekly' and 'monthly' fields under 'periods', 'member' and 'bulk' objects in the API OpenAPI doc were consolidated into a single 'period' field. This change allows for greater flexibility and simplicity when handling different types of periods and aligns with our continuous effort to streamline our API design.