Add 'illusory-doc-logo-light.svg' and update logo paths in 'mint.json'
Added the light version of the Illusory logo as an svg file and incorporated it into the project. Updated the paths of both the light and dark versions of the logo in 'mint.json'. The logos were previously incorrectly referenced, which would lead to display issues. Made this change to correctly display the proper logo variant based on the theme of the documentation page.